pyproject.toml
[project]
name = "hust-bus-stop"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pyspark>=3.5.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
